pyproject.toml
[project]
name = "agent_testbed"
version = "0.1.0"
description = "Agent Testbed"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "google-adk==2.0.0a2",
    "mcp>=1.26.0",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-otlp-proto-grpc",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-instrumentation-google-genai>=0.7b0",
    "opentelemetry-instrumentation-httpx>=0.59b0",
    "opentelemetry-instrumentation-requests>=0.59b0",
    "opentelemetry-instrumentation-vertexai>=2.0b0",
    "opentelemetry-sdk>=1.38.0",
    "uvicorn>=0.41.0",
    "pytest",
    "pytest-asyncio",
    "absl-py>=2.4.0",
    "python-dotenv>=1.2.2",
    "google-auth",
    "google-api-python-client",
    "grpcio",
    "requests",
    "opentelemetry-exporter-gcp-logging>=1.11.0a0",
    "opentelemetry-exporter-gcp-monitoring>=1.11.0a0",
]

[project.scripts]
deploy = "scripts.deploy:main"
run-all = "scripts.run_all:main"
deploy-agent-engine = "scripts.deploy_agent_engine:main"
test-local = "scripts.test_runner:local"
test-remote = "scripts.test_runner:remote"
traffic-loop = "scripts.traffic_loop:main"
verify-traces = "scripts.test_runner:verify_traces"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.uv.workspace]
members = [
    "agents/RootRouter",
    "agents/BookingOrchestrator",
    "agents/CarRentalSpecialist",
    "agents/FlightSpecialist",
    "agents/HotelSpecialist",
    "agents/WeatherSpecialist",
    "mcp_servers/*",
    "traffic_generator"
]

[tool.hatch.build.targets.wheel]
packages = ["scripts", "testbed_utils"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E402", "E501"]

[dependency-groups]
dev = [
    "ruff>=0.15.10",
]
